Move plumbing and config into internal
