+++ cmd/yb/remote_build.go
@@ -22,8 +22,8 @@ import (
 	"github.com/gobwas/ws/wsutil"
 	"github.com/spf13/cobra"
 	"github.com/yourbase/yb"
-	ybconfig "github.com/yourbase/yb/config"
-	"github.com/yourbase/yb/plumbing"
+	ybconfig "github.com/yourbase/yb/internal/config"
+	"github.com/yourbase/yb/internal/plumbing"
 	"gopkg.in/src-d/go-git.v4"
 	gitplumbing "gopkg.in/src-d/go-git.v4/plumbing"
 	"gopkg.in/src-d/go-git.v4/plumbing/object"
